requirements.txt
Pillow~=10.4.0
numpy~=2.0.1
scikit-learn~=1.5.1
matplotlib
pandas~=2.2.2
colormath~=3.0.0